# Add in basic support for JSON generation in BigDataGen and improve performance of from_json