This depends on work done in https://github.com/rapidsai/cudf/pull/14954

In my testing using from_json to pull one column out of a JSON string with only one column in it (10 GiB of data) went from 13.8 seconds to 7.0 seconds. Note that this included reading the 10 GiB from a parquet file too so the speedup is even more.

For longer strings with 512 columns in it and only reading one of the columns (still 10 GiB of data)  went from 17.6 seconds to 11.4 seconds.